Update data classification for Author and Address fields
Modified the data classification annotations for the `Author` property in the `Document` model to use "Yazar" with a sensitivity rank of "Medium" instead of "Author" with a rank of "Low". Updated the `Address` property in the `Home` model to reflect the information type "Home Address" and a rank of "Medium".

Created a new migration class `DocumentAuthor` to manage these schema changes, including the addition and removal of data classification operations for both fields. Updated the `BuildTargetModel` method in the migration designer to incorporate these changes.